handles methods
methods will become a central part in the translation: translating a thread
(resp. a standalone java app) to a Timed Automata, will be reduced to translating
the run() method (resp. the static main() method) to a TA.

The method-oriented translation also eases the task of translating workflows,
because usually workflows are implemented in a method (that invokes others, if
needed).

Important: the method-oriented/workflow translation will start identifying the
relevant attributes/variables to which the method has access, in order to
identify the state that the method has access to. Based on those variables, we
will build the set of TA locations