bugfix in handling preconditions
preconditions are used to add constraints on variables taken from
if/while guards. they can lead to buggy smt models, though, e.g. in the
following example:

if (foo > 0) {
    foo--;
    a++;
}

In this case, when processing instruction a++ we get that:

...
(assert (> foo 0)) ; from the precondition/guard
...
(assert (= foo 0)) ; from a possible source state
...
(assert (= a_1 (+ a 1)) ; from the a++ instruction

the foo conditions form an unsatisfiable core